+++ src/physics_game/Expirable.java
@@ -1,6 +1,6 @@
 package physics_game;
 
 public interface Expirable {
-	public byte getEntityId();
+	public int getEntityId();
 	public boolean isExpired();
 }
